+++ src/com/intellij/advancedExpressionFolding/processor/lombok/LombokFoldingAnnotation.kt
@@ -13,6 +13,7 @@ enum class LombokFoldingAnnotation(val annotation: String) {
     LOMBOK_HASHCODE("@HashCode"),
     LOMBOK_TO_STRING("@ToString"),
     LOMBOK_BUILDER("@Builder"),
+    LOMBOK_UTILITY_CLASS("@UtilityClass"),
 
     LOMBOK_NO_ARGS_CONSTRUCTOR("@NoArgsConstructor"),
     LOMBOK_ALL_ARGS_CONSTRUCTOR("@AllArgsConstructor"),